files sorted for home component - reformat code
