+++ app/src/main/java/com/alphaomardiallo/parisforkids/common/data/repository/queFaireAParis/QueFaireAParisRepositoryImp.kt
@@ -1,7 +1,7 @@
 package com.alphaomardiallo.parisforkids.common.data.repository.queFaireAParis
 
-import com.alphaomardiallo.parisforkids.common.data.remote.RetrofitQueFaireAParis
 import com.alphaomardiallo.parisforkids.common.data.model.responseQueFaireAParis.ResponseQueFaireAParis
+import com.alphaomardiallo.parisforkids.common.data.remote.RetrofitQueFaireAParis
 import retrofit2.Response
 import javax.inject.Inject
